+++ .eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+  extends: 'airbnb',
+  rules: {
+    'no-console': ['error', { allow: ['error'] }],
+    'max-len': ['error', 150],
+  },
+  parser: 'babel-eslint',
+};